Refactor calibration functionality: remove CalibrationWidget from main UI, implement new CalibrationWidget in calibration_tab.py, and update UI elements for better usability.
