+++ src/napari_phasors/_widget.py
@@ -369,170 +369,6 @@ class SdtWidget(AdvancedOptionsWidget):
         super()._on_click(path, reader_options, harmonics)
 
 
-class CalibrationWidget(QWidget):
-    """Widget to calibrate a FLIM image layer using a calibration image."""
-
-    def __init__(self, viewer: "napari.viewer.Viewer"):
-        super().__init__()
-        self.viewer = viewer
-
-        # Creates and empty widget
-        self.calibration_widget = QWidget()
-        uic.loadUi(
-            Path(__file__).parent / "ui/calibration_widget.ui",
-            self.calibration_widget,
-        )
-
-        # Connect callbacks
-        self.calibration_widget.calibrate_push_button.clicked.connect(
-            self._on_click
-        )
-
-        # Connect comboboxes
-        self.calibration_widget.sample_layer_combobox.currentIndexChanged.connect(
-            self._on_combobox_change
-        )
-
-        # Connect layer events to populate combobox
-        self.viewer.layers.events.inserted.connect(self._populate_comboboxes)
-        self.viewer.layers.events.removed.connect(self._populate_comboboxes)
-
-        # Populate combobox
-        self._populate_comboboxes()
-
-        mainLayout = QVBoxLayout()
-        mainLayout.addWidget(self.calibration_widget)
-        self.setLayout(mainLayout)
-
-    def _populate_comboboxes(self):
-        self.calibration_widget.calibration_layer_combobox.clear()
-        image_layers = [
-            layer for layer in self.viewer.layers if isinstance(layer, Image)
-        ]
-        for layer in image_layers:
-            self.calibration_widget.calibration_layer_combobox.addItem(
-                layer.name
-            )
-        self.calibration_widget.sample_layer_combobox.clear()
-        image_layers = [
-            layer for layer in self.viewer.layers if isinstance(layer, Image)
-        ]
-        for layer in image_layers:
-            self.calibration_widget.sample_layer_combobox.addItem(layer.name)
-
-    def _on_combobox_change(self):
-        layer_name = (
-            self.calibration_widget.sample_layer_combobox.currentText()
-        )
-        if layer_name == "":
-            return
-        layer_metadata = self.viewer.layers[layer_name].metadata
-        if (
-            'settings' in layer_metadata.keys()
-            and 'frequency' in layer_metadata['settings'].keys()
-        ):
-            self.calibration_widget.frequency_line_edit_widget.setText(
-                str(layer_metadata['settings']['frequency'])
-            )
-
-    def _on_click(self):
-        sample_name = (
-            self.calibration_widget.sample_layer_combobox.currentText()
-        )
-        calibration_name = (
-            self.calibration_widget.calibration_layer_combobox.currentText()
-        )
-        if sample_name == "" or calibration_name == "":
-            show_error("Select sample and calibration layers")
-            return
-        frequency = self.calibration_widget.frequency_line_edit_widget.text()
-        lifetime = self.calibration_widget.lifetime_line_edit_widget.text()
-        if frequency == "":
-            show_error("Enter frequency")
-            return
-        if lifetime == "":
-            show_error("Enter reference lifetime")
-            return
-        frequency = float(frequency)
-        if "settings" not in self.viewer.layers[sample_name].metadata.keys():
-            self.viewer.layers[sample_name].metadata["settings"] = {}
-        self.viewer.layers[sample_name].metadata["settings"][
-            "frequency"
-        ] = frequency
-        lifetime = float(lifetime)
-        sample_metadata = self.viewer.layers[sample_name].metadata
-        sample_phasor_data = sample_metadata[
-            "phasor_features_labels_layer"
-        ].features
-        harmonics = np.unique(sample_phasor_data["harmonic"])
-        original_mean_shape = sample_metadata["original_mean"].shape
-        calibration_phasor_data = (
-            self.viewer.layers[calibration_name]
-            .metadata["phasor_features_labels_layer"]
-            .features
-        )
-        calibration_mean = self.viewer.layers[calibration_name].metadata[
-            "original_mean"
-        ]
-        calibration_harmonics = np.unique(calibration_phasor_data["harmonic"])
-        if not np.array_equal(harmonics, calibration_harmonics):
-            show_error(
-                "Harmonics in sample and calibration layers do not match"
-            )
-            return
-        calibration_g = np.reshape(
-            calibration_phasor_data["G_original"],
-            (len(calibration_harmonics),) + calibration_mean.shape,
-        )
-        calibration_s = np.reshape(
-            calibration_phasor_data["S_original"],
-            (len(calibration_harmonics),) + calibration_mean.shape,
-        )
-        if "settings" not in sample_metadata.keys():
-            sample_metadata["settings"] = {}
-        if (
-            "calibrated" not in sample_metadata["settings"].keys()
-            or sample_metadata["settings"]["calibrated"] is False
-        ):
-            real, imag = phasor_calibrate(
-                np.reshape(
-                    sample_phasor_data["G_original"],
-                    (len(harmonics),) + original_mean_shape,
-                ),
-                np.reshape(
-                    sample_phasor_data["S_original"],
-                    (len(harmonics),) + original_mean_shape,
-                ),
-                calibration_mean,
-                calibration_g,
-                calibration_s,
-                frequency=frequency,
-                lifetime=lifetime,
-                harmonic=harmonics.tolist(),
-            )
-            sample_phasor_data["G_original"] = real.flatten()
-            sample_phasor_data["S_original"] = imag.flatten()
-            sample_metadata["settings"]["calibrated"] = True
-            calibration_phase, calibration_modulation = (
-                polar_from_reference_phasor(
-                    *phasor_center(
-                        calibration_mean, calibration_g, calibration_s
-                    )[1:],
-                    *phasor_from_lifetime(frequency, lifetime),
-                )
-            )
-            sample_metadata["settings"]["calibration_phase"] = float(
-                calibration_phase[0]
-            )
-            sample_metadata["settings"]["calibration_modulation"] = float(
-                calibration_modulation[0]
-            )
-            print(sample_metadata["settings"])
-            show_info(f"Calibrated {sample_name}")
-        elif sample_metadata["settings"]["calibrated"] is True:
-            show_error("Layer already calibrated")
-
-
 class WriterWidget(QWidget):
     """Widget to export phasor data to a OME-TIF or CSV file."""
